fix: Option usedefaults re-assign
```js
var options = {
  reverse: true,
  usedefaults: false
}
var formatter = new StringMask('+00 (00) 0000-0000')
```
Old code;

```js
usedefaults: this.options.usedefaults || this.options.reverse
```

OR operator (||) return false on left side and return true on right side.

OR for defaults can be convenient because it is short and easy to write but to booleans parameters it was not good.